bpf: Move egress gateway code to its own lib/ file
