+++ bpf/bpf_lxc.c
@@ -34,6 +34,7 @@
 #include "lib/dbg.h"
 #include "lib/trace.h"
 #include "lib/csum.h"
+#include "lib/egress_policies.h"
 #include "lib/encap.h"
 #include "lib/eps.h"
 #include "lib/nat.h"
